Finished full scoring input through GUI
